% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ResIN.R
\name{as.networkx}
\alias{as.networkx}
\title{Python NetworkX coercion helpers}
\usage{
as.networkx(x, ...)
}
\arguments{
\item{x}{Object to coerce/export.}

\item{...}{Passed to methods.}
}
\description{
Generic for exporting objects to a lightweight NetworkX-ready representation.
For \code{ResIN} objects, this creates edge and node CSV tables that can be
read via \pkg{pandas} and converted to a \pkg{networkx} graph.
}
