% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ResIN.R
\name{autoplot.ResIN}
\alias{autoplot.ResIN}
\title{Autoplot for ResIN objects}
\usage{
\method{autoplot}{ResIN}(object, ...)
}
\arguments{
\item{object}{A ResIN object.}

\item{...}{Passed on (currently ignored).}
}
\value{
A \code{ggplot} object.
}
\description{
Produces the ggplot stored in \code{$ResIN_ggplot}.
}
