% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boostrap_example.R
\docType{data}
\name{Bootstrap_example}
\alias{Bootstrap_example}
\title{Output example for a bootstrapping analysis conducted with the ResIN package}
\format{
An object of class \code{"rda"}
}
\usage{
data(Bootstrap_example)
}
\description{
Output example for a bootstrapping analysis conducted with the ResIN package
}
\examples{

data(Bootstrap_example)
names(Bootstrap_example[[1]])

}
\references{
This dataset was made available by Lüders et.al. 2024.
}
\keyword{datasets}
