% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrJSocPsychol_2024.R
\docType{data}
\name{BrJSocPsychol_2024}
\alias{BrJSocPsychol_2024}
\title{Source data for Lüders, A., Carpentras, D. and Quayle, M., 2024. Attitude networks as intergroup realities: Using network‐modelling to research attitude‐identity relationships in polarized political contexts. British Journal of Social Psychology, 63(1), pp.37-51.}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(BrJSocPsychol_2024)
}
\description{
The sample of N = 402 paid participants through the crowd working platform Prolific Academic. The core of the survey consists of A set of eight political attitude items abortion, immigration, gun control, and gay marriage. Each item followed a 5-point scale ranging from strong disagreement to strong agreement. The survey also includes items on partisanship, affective polarization, and a short vignette experiment.
}
\examples{

data(BrJSocPsychol_2024)
head(BrJSocPsychol_2024)


}
\references{
This dataset was made available by Lüders et.al. 2024.
}
\keyword{datasets}
