% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik_data.R
\docType{data}
\name{lik_data}
\alias{lik_data}
\title{Likert-type simulated data for "ResIN" package examples}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(lik_data)
}
\description{
An artificially created data-set (n=1000) of 12, 5-point Likert data. Modeled on the basis of a standard normal data-generating process. Likert scales contain 20 percent uncorrelated, homoscedastic measurement error. This data-set is used for the examples in the "ResIN" package vignette.
}
\examples{

data(lik_data)
head(lik_data)


}
\references{
This data set was artificially created for the ResIN package.
}
\keyword{datasets}
