% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ResIN.R
\name{print.ResIN}
\alias{print.ResIN}
\title{ResIN object S3 methods}
\usage{
\method{print}{ResIN}(x, ...)
}
\arguments{
\item{x}{A ResIN object.}

\item{...}{Ignored.}
}
\description{
ResIN object S3 methods
}
