% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion-generics.R
\name{ResIN-coercion-generics}
\alias{ResIN-coercion-generics}
\alias{as.qgraph}
\alias{as.gephi}
\title{Internal coercion generics}
\usage{
as.qgraph(x, ...)

as.gephi(x, ...)
}
\description{
These functions are S3 generics used for object coercion in \pkg{ResIN}.
End users should typically consult the class methods:
\code{\link{as.igraph.ResIN}}, \code{\link{as.qgraph.ResIN}}, and \code{\link{as.gephi.ResIN}}.
}
\keyword{internal}
