% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{ResIN-reexports}
\alias{ResIN-reexports}
\alias{as.igraph}
\alias{as.network}
\title{Re-exported functions used by ResIN}
\usage{
as.igraph(x, ...)

as.network(x, ...)
}
\description{
These functions are re-exported so users can call common coercion generics
(e.g., \code{as.igraph()} and \code{as.network()}) directly after loading
\pkg{ResIN}, with S3 dispatch to \code{ResIN} methods.
}
\keyword{internal}
