% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ResIN.R
\name{summary.ResIN}
\alias{summary.ResIN}
\title{Summarize a ResIN object}
\usage{
\method{summary}{ResIN}(object, ...)
}
\arguments{
\item{object}{A ResIN object.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary.ResIN}.
}
\description{
Summarize a ResIN object
}
