% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ResIN.R
\name{print.ResIN}
\alias{print.ResIN}
\title{Print a ResIN object}
\usage{
\method{print}{ResIN}(x, ...)
}
\arguments{
\item{x}{A ResIN object.}

\item{...}{Ignored.}
}
\description{
Print a ResIN object
}
