% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{ResIN-reexports}
\alias{ResIN-reexports}
\alias{as.igraph}
\title{Re-exported functions (resolving igraph conflict)}
\usage{
as.igraph(x, ...)
}
\description{
Re-exported functions (resolving igraph conflict)
}
\keyword{internal}
